feat(review_page): change in get api , add post api and update UI

+++ make_my_trip/lib/features/review/data/model/review_model.dart
@@ -1,4 +1,5 @@
 import 'dart:convert';
+
 /// _id : 1
 /// user_id : {"user_name":"Rushi Gandhi hhh","user_image":"https://cdn.pixabay.com/photo/2020/07/01/12/58/icon-5359553_960_720.png"}
 /// hotel_id : 1
@@ -10,20 +11,24 @@ import 'dart:convert';
 /// location : 2
 /// facilities : 1
 
-ReviewModel reviewModelFromJson(String str) => ReviewModel.fromJson(json.decode(str));
+ReviewModel reviewModelFromJson(String str) =>
+    ReviewModel.fromJson(json.decode(str));
+
 String reviewModelToJson(ReviewModel data) => json.encode(data.toJson());
+
 class ReviewModel {
   ReviewModel({
-      int? id, 
-      UserId? userId, 
-      int? hotelId, 
-      String? date, 
-      double? rating, 
-      String? comment, 
-      int? cleanliness,
-      int? comfort,
-      int? location,
-      int? facilities,}){
+    int? id,
+    UserId? userId,
+    int? hotelId,
+    String? date,
+    double? rating,
+    String? comment,
+    double? cleanliness,
+    double? comfort,
+    double? location,
+    double? facilities,
+  }) {
     _id = id;
     _userId = userId;
     _hotelId = hotelId;
@@ -34,61 +39,76 @@ class ReviewModel {
     _comfort = comfort;
     _location = location;
     _facilities = facilities;
-}
+  }
 
   ReviewModel.fromJson(dynamic json) {
     _id = json['_id'];
     _userId = json['user_id'] != null ? UserId.fromJson(json['user_id']) : null;
     _hotelId = json['hotel_id'];
     _date = json['date'];
-    _rating = json['rating'];
+    _rating = json['rating'].toDouble();
     _comment = json['comment'];
-    _cleanliness = json['cleanliness'];
-    _comfort = json['comfort'];
-    _location = json['location'];
-    _facilities = json['facilities'];
+    _cleanliness = json['cleanliness'].toDouble();
+    _comfort = json['comfort'].toDouble();
+    _location = json['location'].toDouble();
+    _facilities = json['facilities'].toDouble();
   }
+
   int? _id;
   UserId? _userId;
   int? _hotelId;
   String? _date;
   double? _rating;
   String? _comment;
-  int? _cleanliness;
-  int? _comfort;
-  int? _location;
-  int? _facilities;
-ReviewModel copyWith({  int? id,
-  UserId? userId,
-  int? hotelId,
-  String? date,
-  double? rating,
-  String? comment,
-  int? cleanliness,
-  int? comfort,
-  int? location,
-  int? facilities,
-}) => ReviewModel(  id: id ?? _id,
-  userId: userId ?? _userId,
-  hotelId: hotelId ?? _hotelId,
-  date: date ?? _date,
-  rating: rating ?? _rating,
-  comment: comment ?? _comment,
-  cleanliness: cleanliness ?? _cleanliness,
-  comfort: comfort ?? _comfort,
-  location: location ?? _location,
-  facilities: facilities ?? _facilities,
-);
+  double? _cleanliness;
+  double? _comfort;
+  double? _location;
+  double? _facilities;
+
+  ReviewModel copyWith({
+    int? id,
+    UserId? userId,
+    int? hotelId,
+    String? date,
+    double? rating,
+    String? comment,
+    double? cleanliness,
+    double? comfort,
+    double? location,
+    double? facilities,
+  }) =>
+      ReviewModel(
+        id: id ?? _id,
+        userId: userId ?? _userId,
+        hotelId: hotelId ?? _hotelId,
+        date: date ?? _date,
+        rating: rating ?? _rating,
+        comment: comment ?? _comment,
+        cleanliness: cleanliness ?? _cleanliness,
+        comfort: comfort ?? _comfort,
+        location: location ?? _location,
+        facilities: facilities ?? _facilities,
+      );
+
   int? get id => _id;
+
   UserId? get userId => _userId;
+
   int? get hotelId => _hotelId;
+
   String? get date => _date;
+
   double? get rating => _rating;
+
   String? get comment => _comment;
-  int? get cleanliness => _cleanliness;
-  int? get comfort => _comfort;
-  int? get location => _location;
-  int? get facilities => _facilities;
+
+  double? get cleanliness => _cleanliness;
+
+  double? get comfort => _comfort;
+
+  double? get location => _location;
+
+  double? get facilities => _facilities;
 
   Map<String, dynamic> toJson() {
     final map = <String, dynamic>{};
@@ -106,34 +126,43 @@ ReviewModel copyWith({  int? id,
     map['facilities'] = _facilities;
     return map;
   }
-
 }
 
 /// user_name : "Rushi Gandhi hhh"
 /// user_image : "https://cdn.pixabay.com/photo/2020/07/01/12/58/icon-5359553_960_720.png"
 
 UserId userIdFromJson(String str) => UserId.fromJson(json.decode(str));
+
 String userIdToJson(UserId data) => json.encode(data.toJson());
+
 class UserId {
   UserId({
-      String? userName, 
-      String? userImage,}){
+    String? userName,
+    String? userImage,
+  }) {
     _userName = userName;
     _userImage = userImage;
-}
+  }
 
   UserId.fromJson(dynamic json) {
     _userName = json['user_name'];
     _userImage = json['user_image'];
   }
+
   String? _userName;
   String? _userImage;
-UserId copyWith({  String? userName,
-  String? userImage,
-}) => UserId(  userName: userName ?? _userName,
-  userImage: userImage ?? _userImage,
-);
+
+  UserId copyWith({
+    String? userName,
+    String? userImage,
+  }) =>
+      UserId(
+        userName: userName ?? _userName,
+        userImage: userImage ?? _userImage,
+      );
+
   String? get userName => _userName;
+
   String? get userImage => _userImage;
 
   Map<String, dynamic> toJson() {
@@ -142,5 +171,4 @@ UserId copyWith({  String? userName,
     map['user_image'] = _userImage;
     return map;
   }
-
 }
